pyproject.toml
[tool.poetry]
name = "fake-records"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.2.1"
pytest-mock = "^3.10.0"
pandas = "^1.5.3"
requests = "^2.28.2"
kaggle = "^1.5.12"
python-dotenv = "^0.21.1"
urllib3 = "^1.26.14"
beautifulsoup4 = "^4.11.2"
lxml = "^4.9.2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.9.0"
ipykernel = "^6.21.2"


[tool.poetry.group.autograder.dependencies]
gitpython = "^3.1.30"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
